core: extract FiberTest from AsyncTest
